last ver libft, all done

+++ 00/libft/ft_memccpy.c
@@ -1,11 +1,11 @@
 #include "libft.h"
 
-void	*ft_memccpy(void *dest, const void *src, int c, int n)
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
 	unsigned char		*d;
 	const unsigned char	*s;
 	unsigned char		letter;
-	int					i;
+	size_t				i;
 
 	d = (unsigned char *)dest;
 	s = (const unsigned char *)src;
